cli/queries: test Deployment.Result
